+++ exonum-java-binding-core/src/test/java/com/exonum/binding/common/message/TemplateRawTransaction.java
@@ -37,5 +37,6 @@ public class TemplateRawTransaction {
   }
 
   private TemplateRawTransaction() {
+    throw new AssertionError("Non-instantiable");
   }
 }